# CAD-4696 Optimise `OnDisk` tests